Add files via upload
Well, this was a part of my Presentation on Use Cases of Big Data in college. I chose a project on Singapore because it give me a wonderful opportunity to learn more about this wonderfully beautiful country. As i proceeded to research, I was all the more impressed particularly with the readiness of the Government in improving the lifestyles of Singaporeans. Almost every facet of the public lifestyle; be it healthcare, infrastructure, cyber-security, transport, education, communication is ably supported by the latest technological developments found around the globe. I was particularly impressed by the Prime Minister, Lee Hsien Loong, and his willingness in making Singapore the World's First smart Nation.